Weapon-pickup-range
Changed the scoring for what weapons the AI should pick up from int to float. This increases the distance at which the AI is willing to look for a dropped weapon on the ground tremendously as the score will no longer become 0 due to rounding.

Note: In this context I noticed that weapon-pickup, as it is currently coded only works for basic-alien-weaponry and not for earthling-stuff or mod-stuff. In a future rewrite I plan to make it much more agile and work with anything.

+++ src/Battlescape/BattlescapeGame.cpp
@@ -2596,7 +2596,7 @@ BattleItem *BattlescapeGame::surveyItems(BattleAction *action, bool pickUpWeapon
 	}
 
 	BattleItem *targetItem = 0;
-	int maxWorth = 0;
+	float maxWorth = 0;
 
 	// now select the most suitable candidate depending on attractiveness and distance
 	// (are we still talking about items?)
@@ -2606,7 +2606,7 @@ BattleItem *BattlescapeGame::surveyItems(BattleAction *action, bool pickUpWeapon
 		{
 			continue;
 		}
-		int currentWorth = bi->getRules()->getAttraction() / ((Position::distance2d(action->actor->getPosition(), bi->getTile()->getPosition()) * 2)+1);
+		float currentWorth = bi->getRules()->getAttraction() / ((Position::distance2d(action->actor->getPosition(), bi->getTile()->getPosition()) * 2) + 1);
 		if (currentWorth > maxWorth)
 		{
 			if (bi->getTile()->getTUCost(O_OBJECT, action->actor->getMovementType()) == 255)
